Share P3M + outdated merge logic across Python package managers
Extract the parallel P3M-fetch / outdated-set merge from
PipPackageManager and UvPackageManager into a single
`fetchMetadataWithOutdated` helper. Each manager now only owns the
tool-specific shell-out (`pip list --outdated` / `uv pip list --outdated`)
and returns a Set of lowercase names; the helper handles the parallel
fetch, error swallowing, and merge into the P3M metadata map.

Net -39 lines.
